requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
scikit-learn>=1.3.0
nltk>=3.8
tqdm>=4.65.0
numpy>=1.24.0
pyyaml>=6.0

